Add the projects in separate file
