+++ src/components/Work/index.js
@@ -1,65 +1,12 @@
 /* eslint-disable */
 import './index.scss'
 import Loader from 'react-loaders'
-import { useEffect, useRef, useState } from 'react';
-import projImg1 from "../../assets/images/project-img1.png";
-import projImg2 from "../../assets/images/project-img2.jpeg";
-import projImg3 from "../../assets/images/project-img3.png";
-import projImg4 from "../../assets/images/project-img4.png";
-import projImg5 from "../../assets/images/project-img5.png";
-import projImg6 from "../../assets/images/project-img6.png";
-import { ProjectCard } from "./ProjectCard";
+import { useEffect, useState } from 'react';
+import projects from "./projectData";
+import { ProjectCard } from "./projectCard";
 import AnimatedLetters from '../AnimatedLetters';
 
-
-
 const Work = () => {
-
-
-  const projects = [
-    {
-      title: "Football Academy website",
-      description: "Football Academy's website, where you can delve into the world of our exceptional staff and discover the extraordinary programs we offer. Our website is meticulously crafted using a blend of cutting-edge technologies such as HTML, CSS, and Javascript, ensuring an immersive and seamless browsing experience. Get ready to embark on a thrilling journey as you explore the diverse talents of our staff and the comprehensive range of programs we have meticulously designed for aspiring football enthusiasts like yourself.",
-      imgUrl: projImg1,
-      githubLink: "https://github.com/Ouail-01/Capstone_Project",
-      projLink: "https://ouail-01.github.io/Capstone_Project/",
-    },
-    {
-      title: "Musify Application",
-      description: "Musify app is a platform that allows users to listen to music, create playlists, discover new music, and customize their listening experience. The app typically offers a vast library of songs from different genres, artists, and albums, and uses algorithms to curate personalized playlists and recommendations for users based on their listening habits. Some music apps may also offer additional features like lyrics, music videos, podcasts, and live streaming. The aim of a music app is to provide users with a seamless and enjoyable listening experience that caters to their music preferences.",
-      imgUrl: projImg2,
-      githubLink: "https://github.com/Ouail-01/Musify_React-Capstone",
-      projLink: "https://ouail-01.github.io/Musify_React-Capstone/",
-    },
-    {
-      title: "Leaderboard",
-      description: "The Leaderboard website is a platform designed to showcase and track scores submitted by various players. It serves as a hub where users can view and compare their achievements with those of other participants. The website offers a user-friendly interface that presents the scores in a visually appealing and organized manner.",
-      imgUrl: projImg3,
-      githubLink: "https://github.com/Ouail-01/Leaderboard_Project",
-      projLink: "https://ouail-01.github.io/Leaderboard_Project/dist/",
-    },
-    {
-      title: "To-do List",
-      description: "To-do List is a tool that helps to organize your day. It simply lists the things that you need to do and allows you to mark them as complete. You will build a simple website that allows for doing that, and you will do it using ES6 and Webpack!",
-      imgUrl: projImg4,
-      githubLink: "https://github.com/Ouail-01/To-Do-List_Project",
-      projLink: "https://ouail-01.github.io/To-Do-List_Project/",
-    },
-    {
-      title: "Space Travelers Hub",
-      description: "Space Traveler's Hub Project is a Single Page App (SPA) implemeted using React and Redux that has 3 pages: The Rockets section displays a list of all available SpaceX rockets. Users can book each rocket by clicking the reservation button or cancel the previously made booking. The Missions section displays a list of current missions along with their brief description and participation status. There is also a button next to each mission that allows users to join the selected mission or leave the mission the user joined earlier.displayed It uses an external API to SpaceX's API to gather data to be displayed on each page. The Dragons section displays a list of all available SpaceX dragons. Users can book each dragon by clicking the reservation button or cancel the previously made booking.",
-      imgUrl: projImg5,
-      githubLink: "https://github.com/Ouail-01/space-travelers-hub",
-      projLink: "https://space-travelers-hub-rw29.onrender.com/rockets",
-    },
-    {
-      title: "Keeping track of hundreds of components",
-      description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
-      imgUrl: projImg6,
-      githubLink: "https://github.com/Ouail-01",
-      projLink: "https://github.com/Ouail-01",
-    },
-  ]
 
   const [letterClass, setLetterClass] = useState('text-animate');
 
@@ -105,9 +52,6 @@ const Work = () => {
                 })
               }
          </div>
-
-
-
       </div>
       <Loader type="pacman" />
     </>
